KMIG-58 : KCLSUPGRAD-91 KCLSUPGRAD036 My Lists are Not Showing UP for Some Patrons
The reason that the lists were not showing up was because they were too large
and cause server errors.  The fix for this was to add paging to the list items.
The paging allowed for smaller queries that would not time out.

Signed-off-by: Kyle Tomita <tomitakyle@gmail.com>
(cherry picked from commit 28319c303dd136b08a3884bef52f67b0f76e5b03)

Signed-off-by: Kyle Tomita <tomitakyle@gmail.com>